Cleanup and fix memory management
